Style: rename login component
